remove buffer_go18
